+++ src/react-components/src/components/03-organisms/Accordion/Accordion.tsx
@@ -1,5 +1,3 @@
-/** Accordion component that shows content on toggle */
-
 import * as React from "react";
 import Button, { ButtonOptions } from "../../01-atoms/Button/Button";
 import bem from "../../../utils/bem";
@@ -8,14 +6,16 @@ import { ButtonTypes } from "../../01-atoms/Button/ButtonTypes";
 export interface AccordionProps {
   id: string;
 
-  /** Used for BEM css convention. */
+  /** Optional modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
   modifiers?: string[];
-  /** Used for BEM css convention. */
+  /** Optional blockName for use with BEM. See how to work with blockNames and BEM here: http://getbem.com/introduction/ */
   blockName?: string;
 
   /** The button that controls the acccordion toggle */
   buttonOptions: ButtonOptions;
 }
+
+/** Accordion component that shows content on toggle */
 
 export default class Accordion extends React.Component<AccordionProps, { isOpen: boolean }> {
   constructor(props: AccordionProps) {