+++ src/react-components/src/components/03-organisms/List/UnorderedList.tsx
@@ -1,4 +1,3 @@
-/** Wrapper component for an unordered list of objects */
 import * as React from "react";
 import bem from "../../../utils/bem";
 import { boolean } from "@storybook/addon-knobs";
@@ -15,11 +14,13 @@ interface UnorderedListProps {
   /** The unique ID of UnorderedList */
   id: string;
 
-  /** Used for BEM css convention. */
+  /** Optional modifiers array for use with BEM. See how to work with modifiers and BEM here: http://getbem.com/introduction/ */
   modifiers?: any[];
-  /** Used for BEM css convention. */
+  /** Optional blockName for use with BEM. See how to work with blockNames and BEM here: http://getbem.com/introduction/ */
   blockName?: string;
 }
+
+/** Wrapper component for an unordered list of objects */
 
 export default function UnorderedList(props: React.PropsWithChildren<UnorderedListProps>) {
   const { id, modifiers = [], blockName } = props;
@@ -35,7 +36,7 @@ export default function UnorderedList(props: React.PropsWithChildren<UnorderedLi
   });
 
   return (
-    <div className={bem("list-container", modifiers, blockName)}>
+    <div className={bem("list-container", modifiers, baseClass)}>
       <ul id={id} className={bem(baseClass)}>
         {content}
       </ul>